Add run_in_container helper and improve copy_to_container()
Adds a new helper run_in_container '...' that executes a command in
the container and commits the resulting image.

copy_to_container sets the access permissions in the container to match
the local file. This allows adding an executable script in a single step
rather than adding it and either invoking it via 'bash ...' or separately
running 'chmod +x ...' on it.