Fix plugin to work with latest version of buildstep.
The latest version of buildstep merges the functionality of /exec and /start.
This caused the prior version of this plugin to enter an infinite fork as the
supervisord config would execute the child processes via /exec.

The post-release step now:
1) Removes /exec if it is a link.
2) Copies /start to /exec (as a file now, not a link)
3) Replaces /start with our custom copy (that runs supervisord)

The custom /start remains the entry point to start the dokku app. It generates
a supervisord config that invokes the Procfile commands via /exec-cmd.

Thanks for Michael Hobbs and Olexandr Sydorchuk for a lot of testing and
a variety of patches for workarounds.