Improve SCALE file support
Improves SCALE file support by allowing the skipping of process types.
If a Procfile entry does not have an entry in the SCALE file, or if
it has an entry with a num processes of 0 (zero), then it will be skipped
and no supervisord config will not be generated for it.

The default behavior of runinng one process for each entry in the Procfile
remains when a SCALE file is not defined.